feat: internal string translation functions
