+++ packages/i18n/src/fallbacks.ts
@@ -1,2 +1,1 @@
-export * from './fallbacks/gtFallback';
-export * from './fallbacks/mFallback';
+export * from './translation-functions/fallbacks';
