+++ packages/i18n/src/index.ts
@@ -1,2 +1,3 @@
-export * from './msg';
+export * from './translation-functions/msg';
 export * from './translation-functions/static';
+export * from './translation-functions/fallbacks';
