--- packages/i18n/src/logs/warnings.ts
@@ -1,1 +1,0 @@
-export const interpolationFailureWarning = `gt-i18n warning: String interpolation failed.`;
