+++ packages/i18n/src/translation-functions/fallbacks/gtFallback.ts
@@ -1,10 +1,5 @@
-import { GTFunctionType, InlineTranslationOptions } from '../../types';
-import logger from '../../logs/logger';
-import { formatMessage } from './utils/formatMessage';
-import { extractVariables } from '../../utils/extractVariables';
-import { interpolationFailureWarning } from '../../logs/warnings';
-import { VAR_IDENTIFIER } from 'generaltranslation/internal';
-import { formatCutoff } from 'generaltranslation';
+import { GTFunctionType, InlineTranslationOptions } from '../types';
+import { interpolateMessage } from '../utils/interpolateMessage';
 
 /**
  * A fallback function for the gt() function that decodes and interpolates.
@@ -35,26 +30,5 @@ export const gtFallback: GTFunctionType = <T extends string | null | undefined>(
   encodedMsg: T,
   options: InlineTranslationOptions = {}
 ): T extends string ? string : T => {
-  // Return if the encoded message is null or undefined
-  if (!encodedMsg) return encodedMsg as T extends string ? string : T;
-
-  // Remove any gt related options
-  const variables = extractVariables(options);
-
-  try {
-    // Interpolate the message
-    const interpolatedMessage = formatMessage(encodedMsg, {
-      ...variables,
-      [VAR_IDENTIFIER]: 'other',
-    });
-    // Apply cutoff formatting
-    const cutoffMessage = formatCutoff(interpolatedMessage, {
-      maxChars: options.$maxChars,
-    });
-    return cutoffMessage as T extends string ? string : T;
-  } catch {
-    // Fallback to decodeMsg
-    logger.warn(interpolationFailureWarning);
-    return encodedMsg as T extends string ? string : T;
-  }
+  return interpolateMessage(encodedMsg, options);
 };
