+++ packages/i18n/src/translation-functions/utils/isEncodedTranslationOptions.ts
@@ -1,4 +1,4 @@
-import { EncodedTranslationOptions } from '../../../types';
+import { EncodedTranslationOptions } from '../../types';
 
 /**
  * Given a decoded options object, validate that includes required decoded options
