+++ packages/i18n/src/translation-functions/utils/messages.ts
@@ -1,0 +1,1 @@
+export const interpolationFailureMessage = `String interpolation failed.`;
